refactor: dto 분리
