+++ backend/src/main/java/com/festago/dto/AdminFestivalResponse.java
@@ -1,0 +1,12 @@
+package com.festago.dto;
+
+import java.time.LocalDate;
+
+public record AdminFestivalResponse(
+    Long id,
+    String name,
+    LocalDate startDate,
+    LocalDate endDate,
+    String thumbnail) {
+
+}
