+++ backend/src/main/java/com/festago/dto/AdminResponse.java
@@ -1,0 +1,10 @@
+package com.festago.dto;
+
+import java.util.List;
+
+public record AdminResponse(
+    List<AdminTicketResponse> adminTickets,
+    List<AdminStageResponse> adminStageResponse,
+    List<AdminFestivalResponse> adminFestivalResponse) {
+
+}
