+++ backend/src/main/java/com/festago/dto/AdminStageResponse.java
@@ -1,0 +1,13 @@
+package com.festago.dto;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+public record AdminStageResponse(
+    Long id,
+    Long festivalId,
+    LocalDateTime startTime,
+    String lineUp,
+    List<Long> ticketId) {
+
+}
